More AuthStatus and AuthTokenServiceImpl into their own files
